LayoutTests/imported/w3c/web-platform-tests/web-animations/timing-model/animations/set-the-animation-start-time-expected.txt


FAIL Setting the start time of an animation without an active timeline assert_equals: Setting the start time clears the current time expected (object) null but got (number) 1000
PASS Setting an unresolved start time an animation without an active timeline does not clear the current time 
FAIL Setting the start time clears the hold time assert_approx_equals: The current time is calculated from the start time, not the hold time expected 2000 +/- 0.0005 but got 1000
FAIL Setting an unresolved start time sets the hold time assert_approx_equals: Hold time is set after start time is made unresolved expected a number but got a "object"
FAIL Setting the start time resolves a pending ready promise assert_true: Animation is in play-pending state expected true got false
FAIL Setting the start time resolves a pending pause task assert_true: Animation is in pause-pending state expected true got false
PASS Setting the start time updates the finished state 
